fixing errors and refactoring

+++ VulkanGraphics/VKBase/vk_descriptors.cpp
@@ -192,7 +192,7 @@ namespace nekographics {
         for (auto& write : writes) {
             write.dstSet = set;
         }
-        vkUpdateDescriptorSets(pool.m_PoolDevice.device(), writes.size(), writes.data(), 0, nullptr);
+        vkUpdateDescriptorSets(pool.m_PoolDevice.device(), static_cast<uint32_t>(writes.size()), writes.data(), 0, nullptr);
     }
 
 }  // namespace lve
